feat: multithreading the solver

+++ include/DatabaseLogger.h
@@ -20,9 +20,9 @@ class DatabaseLogger
 	private:
 		sqlite3* db = nullptr;
 		sqlite3_stmt* stmt = nullptr;
-		int version;
+		double version;
 	public:
-		DatabaseLogger(const string& path, int version);
+		DatabaseLogger(const string& path, double version);
 		~DatabaseLogger();
 	
 		int sqlite3_log_db(vector<DBCube>& cubes);
@@ -32,7 +32,7 @@ class DatabaseLogger
 		void sqlite3_load(vector<DBCube>& data, string& query);
 		void sqlite3_load_version(vector<DBCube>& data);
 		
-		void sqlite3_delete_by_version(int version);
+		void sqlite3_delete_by_version(double version);
 		
-		int sqlite3_count_by_version(int version);
+		int sqlite3_count_by_version(double version);
 };
